Do not restart the server after every converge.
There is a serious defficieny in the upsteam minecraft cookbook in
which it writes configs and sends service resart notifications, which
results in the minecraft server rewriting configs, which in turn causes
chef to detect changes and rewrite the config and the cookbook restarts
the service -- repeat endlessly.

https://github.com/gregf/chef-minecraft#known-issues

The way this _should_ be handled is to load the configs into hashes and
compair the ondisk values with that of the cookbook defined attributes.
If they mismatch, then the config should be re-written and a service
restart issued. However, this is just a minecraft server at the end of
the day... and I don't know that I care enough to write a semi-complex
resource to handle that correctly right now. I also am currently
contemplating using spigot instead.